+++ django_fido/models.py
@@ -73,6 +73,7 @@ class Authenticator(models.Model):
     """
 
     user = models.ForeignKey(settings.AUTH_USER_MODEL, related_name='authenticators', on_delete=models.CASCADE)
+    user_handle = models.TextField(unique=True, blank=True)
     create_datetime = models.DateTimeField(auto_now_add=True)
 
     credential_id_data = models.TextField(unique=True)